+++ src/main/java/com/example/coconote/api/channel/channelMember/service/ChannelMemberService.java
@@ -1,6 +1,7 @@
 package com.example.coconote.api.channel.channelMember.service;
 
 import com.example.coconote.api.channel.channel.entity.Channel;
+import com.example.coconote.api.channel.channel.entity.ChannelType;
 import com.example.coconote.api.channel.channel.repository.ChannelRepository;
 import com.example.coconote.api.channel.channelMember.dto.response.ChannelMemberListResDto;
 import com.example.coconote.api.channel.channelMember.entity.ChannelMember;
@@ -8,12 +9,9 @@ import com.example.coconote.api.channel.channelMember.entity.ChannelRole;
 import com.example.coconote.api.channel.channelMember.repository.ChannelMemberRepository;
 import com.example.coconote.api.member.entity.Member;
 import com.example.coconote.api.member.repository.MemberRepository;
-import com.example.coconote.api.section.entity.Section;
 import com.example.coconote.api.section.repository.SectionRepository;
-import com.example.coconote.api.workspace.workspace.entity.Workspace;
 import com.example.coconote.api.workspace.workspace.repository.WorkspaceRepository;
 import com.example.coconote.api.workspace.workspaceMember.entity.WorkspaceMember;
-import com.example.coconote.api.workspace.workspaceMember.entity.WsRole;
 import com.example.coconote.api.workspace.workspaceMember.repository.WorkspaceMemberRepository;
 import com.example.coconote.common.IsDeleted;
 import jakarta.persistence.EntityNotFoundException;
@@ -32,8 +30,6 @@ public class ChannelMemberService {
     private final ChannelMemberRepository channelMemberRepository;
     private final WorkspaceMemberRepository workspaceMemberRepository;
     private final ChannelRepository channelRepository;
-    private final SectionRepository sectionRepository;
-    private final WorkspaceRepository workspaceRepository;
     private final MemberRepository memberRepository;
 
     @Autowired
@@ -46,8 +42,6 @@ public class ChannelMemberService {
         this.channelMemberRepository = channelMemberRepository;
         this.workspaceMemberRepository = workspaceMemberRepository;
         this.channelRepository = channelRepository;
-        this.sectionRepository = sectionRepository;
-        this.workspaceRepository = workspaceRepository;
         this.memberRepository = memberRepository;
     }
 
@@ -114,6 +108,9 @@ public class ChannelMemberService {
         if(deleter.getChannelRole().equals(ChannelRole.USER)) {
             throw new IllegalArgumentException("회원을 강퇴시킬 권한이 없습니다.");
         }
+        if(channel.getChannelType().equals(ChannelType.DEFAULT)) {
+            throw new IllegalArgumentException("기본 채널에서는 다른 회원을 퇴장시킬 수 없습니다.");
+        }
         channelMember.deleteEntity();
     }
 
@@ -153,4 +150,14 @@ public class ChannelMemberService {
         return channelMember.fromEntity();
     }
 
+    public void channelMemberLeave(Long channelId, String email) {
+        Member member = memberRepository.findByEmail(email).orElseThrow(()-> new EntityNotFoundException("회원을 조회할 수 없습니다."));
+        Channel channel = channelRepository.findById(channelId).orElseThrow(()-> new EntityNotFoundException("채널을 조회할 수 없습니다."));
+        WorkspaceMember workspaceMember = workspaceMemberRepository.findByMemberAndWorkspaceAndIsDeleted(member, channel.getSection().getWorkspace(), IsDeleted.N).orElseThrow(()-> new EntityNotFoundException("존재하지 않는 워크스페이스 회원입니다."));
+        ChannelMember channelMember = channelMemberRepository.findByChannelAndWorkspaceMemberAndIsDeleted(channel, workspaceMember, IsDeleted.N).orElseThrow(()-> new EntityNotFoundException("채널 회원을 조회할 수 없습니다."));
+        if(channel.getChannelType().equals(ChannelType.DEFAULT)) {
+            throw new IllegalArgumentException("기본 채널에서는 나갈 수 없습니다.");
+        }
+        channelMember.deleteEntity();
+    }
 }
